Correction d'erreur + Structure des section

+++ Alone/PathTracer_bitmap.cpp
@@ -40,7 +40,7 @@ namespace PathTracerNS
 	BYTE* LoadBMP ( int* width, int* height, long* size, char* filePath )
 	{
 		// Creation of the LPCTSTR bmpfile
-		int cubeMapPathLength = strlen( filePath );
+		int cubeMapPathLength = (int) strlen( filePath );
 		wchar_t *cubeMapPathWChr = new wchar_t[cubeMapPathLength];
 		MultiByteToWideChar(0, 0, filePath, cubeMapPathLength + 1, cubeMapPathWChr, cubeMapPathLength + 1);
 		LPCTSTR bmpfile = cubeMapPathWChr;
